add custom base command
https://coveord.atlassian.net/browse/CDX-348
